+++ release/ui/typography/Section/Section.d.ts
@@ -1,7 +1,11 @@
 import { type StateProp } from '@innet/dom';
 import { type FlexProps } from '../../layout';
 import { type TitleProps } from '../Title';
-export interface SectionProps extends FlexProps<HTMLDivElement> {
+export interface SectionStyles {
+    root: any;
+    title: any;
+}
+export interface SectionProps extends FlexProps<HTMLDivElement, {}, SectionStyles> {
     title?: string;
     subTitle?: StateProp<string>;
     titleProps?: TitleProps;